requirements.txt
openai
python-dotenv
edge-tts
moviepy
requests
pillow
fastapi
uvicorn
pydantic
google-generativeai
